feat: remove unused a11y examples and refact budget
