+++ src/components/resume/a11y-practice/A11yPractice.test.tsx
@@ -15,7 +15,9 @@ describe("resume | a11y-practice | A11yPractice", () => {
     expect(
       screen.getByRole("tab", { name: "Chunking", selected: true }),
     ).toBeInTheDocument();
-    expect(screen.getByRole("tab", { name: "+ Debounce" })).toBeInTheDocument();
+    expect(
+      screen.getByRole("tab", { name: "Chunking + Debounce" }),
+    ).toBeInTheDocument();
   });
 
   it("shows the v1 stream example on the default tab", () => {
@@ -30,10 +32,13 @@ describe("resume | a11y-practice | A11yPractice", () => {
   it("switches to the debounce tab when selected", () => {
     render(<A11yPractice />);
 
-    fireEvent.click(screen.getByRole("tab", { name: "+ Debounce" }));
+    fireEvent.click(screen.getByRole("tab", { name: "Chunking + Debounce" }));
 
     expect(
-      screen.getByRole("tab", { name: "+ Debounce", selected: true }),
+      screen.getByRole("tab", {
+        name: "Chunking + Debounce",
+        selected: true,
+      }),
     ).toBeInTheDocument();
     expect(
       screen.getByRole("button", { name: /start stream/i }),
@@ -44,6 +49,6 @@ describe("resume | a11y-practice | A11yPractice", () => {
     render(<A11yPractice />);
 
     const panels = screen.getAllByRole("tabpanel", { hidden: true });
-    expect(panels.length).toBe(10);
+    expect(panels.length).toBe(2);
   });
 });
