+++ src/components/resume/finances/budgeting/entries/ExpenseEntryDialog.tsx
@@ -21,12 +21,16 @@ import {
   Typography,
 } from "@mui/material";
 import {
+  formatCategoryName,
+  getPercentSources,
+  getTaxBasis,
+} from "../../../../../apis/budget";
+import {
   ExpenseEntry,
   ExpensePercentSource,
   ExpenseTaxBasis,
   ExpenseValueMode,
-} from "../../../../jotai/finances-atom";
-import { getPercentSources, formatCategoryName, getTaxBasis } from "./helpers";
+} from "../../../../../jotai/finances-atom";
 
 const tfProps: TextFieldProps = {
   variant: "standard",
