--- src/components/resume/finances/budgeting/helpers/index.ts
@@ -1,2 +1,0 @@
-export * from "../../../../../apis/budget";
-export * from "../graphs/chartData";
